tecnologias-web/projeto-jpa-dto-query-com-tratamento-excecoes/projeto-jpa-dto-query/src/main/java/sptech/projetojpadtoquery/controle/MotoristaController.java
package sptech.projetojpadtoquery.controle;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import sptech.projetojpadtoquery.dominio.Motorista;
import sptech.projetojpadtoquery.repositorio.MotoristaRepository;
import sptech.projetojpadtoquery.resposta.MotoristaSimplesResponse;

import java.util.List;

@RestController
@RequestMapping("/motoristas")
public class MotoristaController {

    @Autowired
    private MotoristaRepository repository;

    @GetMapping
    public ResponseEntity<List<Motorista>> get(){
        List<Motorista> lista = repository.findAll();
        if(lista.isEmpty()){
            return ResponseEntity.status(204).build();
        }
        return ResponseEntity.status(200).body(lista);
    }

    @GetMapping("/simples")
    public ResponseEntity<List<MotoristaSimplesResponse>> getSimples() {
        List<MotoristaSimplesResponse> lista = repository.getMotoristasSimples();
        if(lista.isEmpty()){
            return ResponseEntity.status(204).build();
        }
        return ResponseEntity.status(200).body(lista);
    }
}
